LAB-2/requirements.txt
streamlit>=1.35.0
groq>=0.5.0
python-dotenv>=1.0.1
sacrebleu